refactor(direct-control): extract proof log helpers
Refs: openspec/changes/civ7-support-direct-control-modularization

Move FileSnapshot, FreshLogMarkerProof, snapshotFile, waitForFreshLogMarkers, and their private ordered-marker/file helpers into src/proof/log-markers.ts while keeping the public package facade exports in index.ts.

This preserves missing-file snapshots, timeout and poll defaults, startOffset/truncated-file behavior, ordered marker matching, rejectPattern handling, and log-timeout classification/details. Capability catalog, operation/proof telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, and Effect/oRPC procedure-core work remain pending under the compatibility matrix.

Verification:
- bun run --cwd packages/civ7-direct-control test -- runtime-and-catalog.test.ts session.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- git diff --check
- proof helper ownership/no-cycle scans

Local package/source relocation proof only; FreshLogMarkerProof is local/log evidence and no runtime or live-game proof is claimed.
