+++ packages/civ7-direct-control/src/index.ts
@@ -1,4 +1,4 @@
-import { open, readdir, readFile, stat } from "node:fs/promises";
+import { readdir, readFile, stat } from "node:fs/promises";
 import { homedir } from "node:os";
 import { basename, extname, join, resolve } from "node:path";
 import { Socket, createConnection } from "node:net";
@@ -26,6 +26,12 @@ import {
   parseCiv7TunerFrame,
   type Civ7TunerFrame,
 } from "./session/framing.js";
+import {
+  snapshotFile,
+  waitForFreshLogMarkers,
+  type FileSnapshot,
+  type FreshLogMarkerProof,
+} from "./proof/log-markers.js";
 import {
   appUiSnapshotFromCommandResult,
   buildAppUiSnapshotCommand,
@@ -96,6 +102,14 @@ export {
   parseCiv7TunerFrame,
 } from "./session/framing.js";
 export type { Civ7TunerFrame } from "./session/framing.js";
+export {
+  snapshotFile,
+  waitForFreshLogMarkers,
+} from "./proof/log-markers.js";
+export type {
+  FileSnapshot,
+  FreshLogMarkerProof,
+} from "./proof/log-markers.js";
 
 export const DEFAULT_CIV7_TUNER_HOST = "127.0.0.1";
 export const DEFAULT_CIV7_TUNER_PORT = 4318;
@@ -4048,112 +4062,6 @@ export async function loadCiv7OfficialResourceCapabilities(options: {
   return Array.from(found.values());
 }
 
-export type FileSnapshot = Readonly<{
-  exists: boolean;
-  size: number;
-  mtimeMs: number;
-  prefix: string;
-  prefixBytes: number;
-}>;
-
-export type FreshLogMarkerProof = Readonly<{
-  logPath: string;
-  observedAt: string;
-  startOffset: number;
-  matched: ReadonlyArray<string>;
-}>;
-
-export async function snapshotFile(path: string): Promise<FileSnapshot> {
-  const info = await stat(path).catch((err: unknown) => {
-    if (isNodeNotFound(err)) return null;
-    throw err;
-  });
-  return info
-    ? {
-        exists: true,
-        size: info.size,
-        mtimeMs: info.mtimeMs,
-        ...(await filePrefixSnapshot(path, info.size)),
-      }
-    : { exists: false, size: 0, mtimeMs: 0, prefix: "", prefixBytes: 0 };
-}
-
-async function filePrefixSnapshot(path: string, size: number): Promise<Pick<FileSnapshot, "prefix" | "prefixBytes">> {
-  const prefixBytes = Math.min(size, 4096);
-  if (prefixBytes === 0) return { prefix: "", prefixBytes: 0 };
-  const handle = await open(path, "r");
-  try {
-    const buffer = Buffer.alloc(prefixBytes);
-    const { bytesRead } = await handle.read(buffer, 0, prefixBytes, 0);
-    return {
-      prefix: buffer.subarray(0, bytesRead).toString("utf8"),
-      prefixBytes: bytesRead,
-    };
-  } finally {
-    await handle.close();
-  }
-}
-
-export function logTextFromSnapshot(args: {
-  fullText: string;
-  snapshot: FileSnapshot;
-  current: FileSnapshot;
-}): { text: string; startOffset: number; rewritten: boolean } {
-  const { fullText, snapshot, current } = args;
-  if (!snapshot.exists) return { text: fullText, startOffset: 0, rewritten: true };
-
-  const rewritten = snapshot.prefixBytes > 0 && !fullText.startsWith(snapshot.prefix);
-  if (rewritten) return { text: fullText, startOffset: 0, rewritten };
-  if (current.size > snapshot.size) {
-    return { text: fullText.slice(snapshot.size), startOffset: snapshot.size, rewritten: false };
-  }
-  if (current.mtimeMs > snapshot.mtimeMs) return { text: fullText, startOffset: 0, rewritten: true };
-  return { text: "", startOffset: snapshot.size, rewritten: false };
-}
-
-export async function waitForFreshLogMarkers(options: {
-  logPath: string;
-  snapshot: FileSnapshot;
-  markers: ReadonlyArray<string>;
-  timeoutMs?: number;
-  pollIntervalMs?: number;
-  rejectPattern?: RegExp;
-}): Promise<FreshLogMarkerProof> {
-  const timeoutMs = options.timeoutMs ?? 90_000;
-  const pollIntervalMs = options.pollIntervalMs ?? 1_000;
-  const startedAt = Date.now();
-  const snapshotOffset = options.snapshot.size;
-  let lastStartOffset = snapshotOffset;
-  let lastError: string | undefined;
-
-  while (Date.now() - startedAt <= timeoutMs) {
-    const current = await snapshotFile(options.logPath);
-    if (current.exists && (current.size > snapshotOffset || current.mtimeMs > options.snapshot.mtimeMs)) {
-      const fullText = await readFile(options.logPath, "utf8");
-      const freshLog = logTextFromSnapshot({ fullText, snapshot: options.snapshot, current });
-      lastStartOffset = freshLog.startOffset;
-      const proof = matchOrderedMarkers(freshLog.text, options.markers);
-      const rejected = options.rejectPattern?.exec(freshLog.text);
-      if (rejected) lastError = `Log contains ${rejected[0]}`;
-      if (proof.ok && !rejected) {
-        return {
-          logPath: options.logPath,
-          observedAt: new Date().toISOString(),
-          startOffset: freshLog.startOffset,
-          matched: proof.matched,
-        };
-      }
-    }
-    await sleep(pollIntervalMs);
-  }
-
-  throw new Civ7DirectControlError(
-    "log-timeout",
-    lastError ?? `Timed out waiting for fresh log markers in ${options.logPath}`,
-    { details: { markers: options.markers, startOffset: lastStartOffset, snapshotOffset } },
-  );
-}
-
 async function openCiv7TunerSocket(options: {
   host: string;
   port: number;
@@ -4271,6 +4179,15 @@ function portFromEnv(env: NodeJS.ProcessEnv): number | undefined {
 
 function splitEnvList(value: string | undefined): string[] {
   return value?.split(",").map((entry) => entry.trim()).filter(Boolean) ?? [];
+}
+
+function isNodeNotFound(err: unknown): boolean {
+  return (
+    err !== null &&
+    typeof err === "object" &&
+    "code" in err &&
+    (err as { code?: unknown }).code === "ENOENT"
+  );
 }
 
 function buildResourcePlacementFeasibilityCommand(input: {
@@ -6521,27 +6438,6 @@ function toDirectControlError(err: unknown, fallbackCode: Civ7DirectControlError
 
 function errorMessage(err: unknown): string {
   return err instanceof Error ? err.message : String(err);
-}
-
-function isNodeNotFound(err: unknown): boolean {
-  return (
-    err !== null &&
-    typeof err === "object" &&
-    "code" in err &&
-    (err as { code?: unknown }).code === "ENOENT"
-  );
-}
-
-function matchOrderedMarkers(text: string, markers: ReadonlyArray<string>): { ok: boolean; matched: string[] } {
-  const matched: string[] = [];
-  let cursor = 0;
-  for (const marker of markers) {
-    const next = text.indexOf(marker, cursor);
-    if (next < 0) return { ok: false, matched };
-    matched.push(marker);
-    cursor = next + marker.length;
-  }
-  return { ok: true, matched };
 }
 
 function sleep(ms: number): Promise<void> {
